requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
pydantic==2.7.1
pydantic-settings==2.3.0
motor==3.4.0
pymongo==4.6.3
httpx==0.27.0
redis==5.0.4
torch==2.3.0
numpy==1.26.4
pandas==2.2.2
scikit-learn==1.5.0
shap==0.45.1
xgboost==2.0.3
scipy==1.13.0
python-dotenv==1.0.1
openmeteo-requests==1.2.0
requests-cache==1.2.1
retry-requests==2.0.0
